GJ - 쇼핑몰 리스트 페이지 만드는중

+++ semi_project/src/main/java/com/web/shoppingmall/model/dto/Product.java
@@ -28,5 +28,5 @@ public class Product {
 	//리뷰 총 갯수
 	private int totalReviewCount;
 	//평균 별점
-	private int avgRating;
+	private double avgRating;
 }
